внесла изменения относительно комментариев
