+++ fixtures/base.js
@@ -18,6 +18,7 @@ import CreateSignatureModal from "../new_pom/modalWindows/createSignatureModal.j
 import FinalStepPage from '../new_pom/pages/finalStepPage.js';
 import SuccessModal from '../new_pom/modalWindows/successModal.js';
 import EditAndResendDocumentModal from "../new_pom/modalWindows/editAndResendDocumentModal.js";
+import NewTemplatesPage from "../new_pom/pages/documents/templatesPage.js";
 
 
 
@@ -124,6 +125,10 @@ export const test = base.extend({
 
     editAndResendDocumentModal: async ({ page }, use) => {
         await use(new EditAndResendDocumentModal(page));
+    },
+
+    templatePage: async ({ page }, use) => {
+        await use(new NewTemplatesPage(page));
     },
