+++ new_pom/pages/sign/signPage.js
@@ -1,8 +1,4 @@
 import SideMenuComponent from "../../components/sideMenuComponent";
-import UploadFileOnSignPage from "../../components/fileUploderComponent";
-import CreateNewTemplateComponent from "../../components/createNewTemplateComponent";
-import SideMenuTemlatesComponent from "../../components/sideMenuTemplatesComponent";
-import NewTemplatesPage from "../documents/templatesPage";
 import UploadFileOnSignPage from "../../pages/sign/uploadFileOnSignPage";
 import HeaderComponent from "../../components/headerComponent";
 
@@ -13,11 +9,6 @@ export default class NewSignPage {
 
     this.uploadFile = new UploadFileOnSignPage(this.page);
     this.sideMenu = new SideMenuComponent(this.page);
-
-
-    this.formNewTemplate = new CreateNewTemplateComponent(this.page);
-    this.sideMenuTemplates = new SideMenuTemlatesComponent(this.page);
-    this.newTemplatesPage = new NewTemplatesPage(this.page);
 
 
     this.header = new HeaderComponent(this.page);
