+++ tests/07.Templates.spec.js
@@ -1,27 +1,24 @@
 import { expect } from "@playwright/test";
 import { test, createBusinessUserAndLogin, signPage } from "../fixtures/base.js";
-// import SideMenuComponent from "../new_pom/components/sideMenuComponent.js";
-// import SideMenuTemlatesComponent from "../new_pom/components/sideMenuTemplatesComponent.js";
-// import NewTemplatesPage from "../new_pom/pages/documents/templatesPage.js";
-// import {} from "../testData.js";
-// import PrepareForSigningModal from "../new_pom/modalWindows/prepareforSigningModal.js";
+
+
 
 
 test.describe('Templates', () => {
 
-  test('TC_07_27_01 | Verify that user can create a template', async ({ createBusinessUserAndLogin, signPage, prepareForSigningModal}) => {
+  test('TC_07_27_01 | Verify that user can create a template', async ({ createBusinessUserAndLogin, signPage, prepareForSignatureModal, templatePage}) => {
 
    await signPage.sideMenu.clickTemplates();
-   await signPage.sideMenuTemplates.clickCreateTemplate();
-   await signPage.formNewTemplate.fillTemplateNameField();
-   await signPage.formNewTemplate.fillOptionalMessageField();
-   await signPage.formNewTemplate.fillCreateTemplateRolesField();
-   await signPage.uploadFile.uploadFile('testDocuments/CSV.csv');
-   await signPage.formNewTemplate.clickfillTemlateBtn();
-   await prepareForSigningModal.clickSignBtn();
-   await prepareForSigningModal.performSignature(100, 100);
-   await prepareForSigningModal.clickcreateBtn();
-   await prepareForSigningModal.clickbackToTempatesBtn();
+   await templatePage.sideMenuTemplates.clickCreateTemplate();
+   await templatePage.formNewTemplate.fillTemplateNameField();
+   await templatePage.formNewTemplate.fillOptionalMessageField();
+   await templatePage.formNewTemplate.fillCreateTemplateRolesField();
+   await signPage.uploadFile.fileUploader.uploadFile('testDocuments/CSV.csv');
+   await templatePage.formNewTemplate.clickfillTemlateBtn();
+   await prepareForSignatureModal.clickSignBtn();
+   await prepareForSignatureModal.performSignature(100, 100);
+   await prepareForSignatureModal.clickcreateBtn();
+   await prepareForSignatureModal.clickbackToTempatesBtn();
 
   });
 });
